Make role re-use easier
The existing proposal provides the foundation needed, but will be somewhat repetitive and prone to lots of boilerplate.

I think it's strongly desirable to make `exec_main` a role meta attribute so you don't have to specify it each time. To avoid the need to do the include-the-other-role dance in a role's tasks, meta role dependencies should be extended to declaratively include the other role. And to make it easier to actually use this, an option to run a different task via the `roles:` stanza in the playbook should be incorporated.

See edits.